Move get_browserhtml_path to command_base #14815
Remove unused import of find_dep_path_newest #14815

Move get_browserhtml_path top along with find_dep_path_newest #14815
